+++ packages/client/src/__tests__/StreamVideoClient.test.ts
@@ -1,9 +1,90 @@
-import { describe, expect, it } from 'vitest';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { StreamVideoClient } from '../StreamVideoClient';
+import 'dotenv/config';
+import { generateUUIDv4 } from '../coordinator/connection/utils';
+import { WSConnectionFallback } from '../coordinator/connection/connection_fallback';
+
+const apiKey = process.env.STREAM_API_KEY!;
+const tokenUrl = process.env.TOKEN_PROVIDER_URL!;
+
+const tokenProvider = (userId: string) => {
+  return async () => {
+    const url = new URL(tokenUrl);
+    url.searchParams.set('api_key', apiKey);
+    url.searchParams.set('user_id', userId);
+
+    const response = await fetch(url.toString());
+    const data = await response.json();
+    return data.token;
+  };
+};
 
 describe('StreamVideoClient', () => {
-  it('can be instantiated', () => {
-    const client = new StreamVideoClient('api-key-123');
-    expect(client).toBeTruthy();
+  let client: StreamVideoClient;
+
+  beforeEach(() => {
+    client = new StreamVideoClient(apiKey, {
+      // tests run in node, so we have to fake being in browser env
+      browser: true,
+    });
+  });
+
+  it('API calls are hold until connection is done', async () => {
+    const user = {
+      id: 'jane',
+    };
+
+    client.connectUser(user, tokenProvider(user.id));
+    const response = await client.queryCalls({});
+
+    expect(response).toBeDefined();
+  });
+
+  it('API calls are hold until connection is done - anonymous users', async () => {
+    client.connectUser({ type: 'anonymous' });
+    const response = await client.queryCalls({
+      filter_conditions: { members: { $in: ['!anon'] } },
+    });
+
+    expect(response).toBeDefined();
+  });
+
+  it('public endpoints can be called without authentication', async () => {
+    const userId = 'guest-' + generateUUIDv4();
+    const response = await client.createGuestUser({
+      user: { id: userId },
+    });
+
+    expect(response.user.id).toContain(userId);
+  });
+
+  it(`private endpoints can't be called without authentication`, async () => {
+    await expect(() => client.queryCalls({})).rejects.toThrowError();
+
+    await client.connectUser({ id: 'jane' }, tokenProvider('jane'));
+    await client.disconnectUser();
+
+    await expect(() => client.queryCalls({})).rejects.toThrowError();
+  });
+
+  it('API calls should be enriched with conneciton id', async () => {
+    const user = {
+      id: 'jane',
+    };
+
+    client.connectUser(user, tokenProvider(user.id));
+    const spy = vi.spyOn(client.streamClient.axiosInstance, 'post');
+    await client.call('default', generateUUIDv4()).getOrCreate();
+
+    const requestConfig: any = spy.mock.calls[spy.mock.calls.length - 1][2];
+    const params = requestConfig.params;
+    expect(client.streamClient._getConnectionID()).toBeDefined();
+    expect(params['connection_id']).toBe(
+      client.streamClient._getConnectionID(),
+    );
+  });
+
+  afterEach(() => {
+    client.disconnectUser();
   });
 });