+++ packages/react-native-sdk/src/providers/MediaStreamManagement/index.tsx
@@ -6,7 +6,11 @@ import React, {
   useMemo,
   useState,
 } from 'react';
-import { CallingState, SfuModels } from '@stream-io/video-client';
+import {
+  CallingState,
+  OwnCapability,
+  SfuModels,
+} from '@stream-io/video-client';
 import {
   useCall,
   useCallCallingState,
@@ -93,11 +97,14 @@ export const MediaStreamManagement = ({ children }: PropsWithChildren<{}>) => {
     useState(true);
 
   const [initAudioEnabled, setInitialAudioEnabled] = useState<boolean>(
-    isMicPermissionGranted$.getValue(),
+    isMicPermissionGranted$.getValue() &&
+      !!call?.permissionsContext?.hasPermission(OwnCapability.SEND_AUDIO),
   );
 
   const [initVideoEnabled, setInitialVideoEnabled] = useState<boolean>(
-    isCameraPermissionGranted$.getValue(),
+    isCameraPermissionGranted$.getValue() &&
+      call?.type !== 'audio_room' &&
+      !!call?.permissionsContext?.hasPermission(OwnCapability.SEND_VIDEO),
   );
 
   const publishVideoStream = useVideoPublisher({