+++ sample-apps/react/react-dogfood/components/MeetingUI.tsx
@@ -35,7 +35,7 @@ import {
   UnreadCountBadge,
 } from '.';
 import { ActiveCallHeader } from './ActiveCallHeader';
-import { useKeyboardShortcuts, useWatchChannel } from '../hooks';
+import { useKeyboardShortcuts, useWakeLock, useWatchChannel } from '../hooks';
 import { DEFAULT_LAYOUT, getLayoutSettings, LayoutMap } from './LayoutSelector';
 import { Stage } from './Stage';
 import { ToggleParticipantListButton } from './ToggleParticipantListButton';
@@ -144,6 +144,7 @@ export const MeetingUI = ({ chatClient, enablePreview }: MeetingUIProps) => {
   }, [activeCall, isSortingDisabled]);
 
   useKeyboardShortcuts();
+  useWakeLock();
 
   let ComponentToRender: JSX.Element | null = null;
   if (show === 'error-join' || show === 'error-leave') {